Changes per review, Final at top-level, doc changes
