+++ package.yaml
@@ -1,5 +1,5 @@
 name:                polysemy
-version:             1.1.0.0
+version:             1.2.0.0
 github:              "isovector/polysemy"
 license:             BSD3
 author:              "Sandy Maguire"
