+++ src/Polysemy/Output.hs
@@ -126,6 +126,14 @@ runOutputMonoidTVar tvar f = interpret $ \case
 -- Internally, this simply creates a new 'IORef', passes it to
 -- 'runOutputMonoidIORef', and then returns the result and the final value
 -- of the 'IORef'.
+--
+-- /Beware/: As this uses an 'IORef' internally,
+-- all other effects will have local
+-- state semantics in regards to 'Output' effects
+-- interpreted this way.
+-- For example, 'Polysemy.Error.throw' and 'Polysemy.Error.catch' will
+-- never revert 'output's, even if 'Polysemy.Error.runError' is used
+-- after 'outputToIOMonoid'.
 outputToIOMonoid
   :: forall o m r a
    . (Monoid m, Member (Embed IO) r)
@@ -146,6 +154,14 @@ outputToIOMonoid f sem = do
 --
 -- You should always use this instead of 'outputToIOMonoid' if the monoid
 -- is a list, such as 'String'.
+--
+-- /Beware/: As this uses an 'IORef' internally,
+-- all other effects will have local
+-- state semantics in regards to 'Output' effects
+-- interpreted this way.
+-- For example, 'Polysemy.Error.throw' and 'Polysemy.Error.catch' will
+-- never revert 'output's, even if 'Polysemy.Error.runError' is used
+-- after 'outputToIOMonoidAssocR'.
 outputToIOMonoidAssocR
   :: forall o m r a
    . (Monoid m, Member (Embed IO) r)
