+++ test/FinalSpec.hs
@@ -3,24 +3,16 @@ module FinalSpec where
 
 import Test.Hspec
 
-import qualified Control.Concurrent.Async as A
-import Control.Concurrent.MVar
-import Control.Concurrent.STM.TVar
-
 import Data.Either
 import Data.IORef
 
 import Polysemy
-import Polysemy.Final
-
-
 import Polysemy.Async
 import Polysemy.Error
 import Polysemy.Fixpoint
 import Polysemy.Trace
 import Polysemy.State
 
-import Polysemy.Writer
 
 data Node a = Node a (IORef (Node a))
 
@@ -84,53 +76,6 @@ test2 =
   trace "Nothing at all."
 
 
-test3 :: IO (String, String)
-test3 = do
-  tvar <- newTVarIO ""
-  (listened, _) <- runFinal . asyncToIOFinal . runWriterTVar tvar $ do
-    tell "message "
-    listen $ do
-      tell "has been"
-      a <- async $ tell " received"
-      await a
-  end <- readTVarIO tvar
-  return (end, listened)
-
-test4 :: IO (String, String)
-test4 = do
-  tvar <- newTVarIO ""
-  lock <- newEmptyMVar
-  (listened, a) <- runFinal . asyncToIOFinal . runWriterTVar tvar $ do
-    tell "message "
-    listen $ do
-      tell "has been"
-      a <- async $ do
-        embedFinal $ takeMVar lock
-        tell " received"
-      return a
-  putMVar lock ()
-  _ <- A.wait a
-  end <- readTVarIO tvar
-  return (end, listened)
-
-test5 :: Sem '[Error (A.Async (Maybe ())), Final IO] String
-test5 = do
-  tvar <- embedFinal $ newTVarIO ""
-  lock <- embedFinal $ newEmptyMVar
-  let
-      inner = do
-        tell "message "
-        fmap snd $ listen @String $ do
-          tell "has been"
-          a <- async $ do
-            embedFinal $ takeMVar lock
-            tell " received"
-          throw a
-  asyncToIOFinal (runWriterTVar tvar inner) `catch` \a ->
-    embedFinal $ do
-      putMVar lock ()
-      (_ :: Maybe ()) <- A.wait a
-      readTVarIO tvar
 
 spec :: Spec
 spec = do
@@ -149,24 +94,3 @@ spec = do
     it "should treat trace with local state semantics" $ do
       res2 <- test2
       res2 `shouldBe` (["Nothing at all."], Right ())
-
-  describe "runWriterTVar" $ do
-    it "should listen and commit asyncs spawned and awaited upon in a listen \
-       \block" $ do
-      (end, listened) <- test3
-      end `shouldBe` "message has been received"
-      listened `shouldBe` "has been received"
-
-    it "should commit writes of asyncs spawned inside a listen block even if \
-       \the block has finished." $ do
-      (end, listened) <- test4
-      end `shouldBe` "message has been received"
-      listened `shouldBe` "has been"
-
-
-    it "should commit writes of asyncs spawned inside a listen block even if \
-       \the block failed for any reason." $ do
-      Right end1 <- runFinal . errorToIOFinal $ test5
-      Right end2 <- runFinal . runError $ test5
-      end1 `shouldBe` "message has been received"
-      end2 `shouldBe` "message has been received"
